% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AverageIntoBins.R
\name{AverageIntoBins}
\alias{AverageIntoBins}
\title{AverageIntoBins}
\usage{
AverageIntoBins(TPMData, logcv, templInfo)
}
\arguments{
\item{TPMData}{numeric matrix, tpmdata}

\item{logcv}{numeric vector, Gene Coefficient of Variation}

\item{templInfo}{list with 5 elements}
}
\value{
a list with cv and meanGeneExpr
}
\description{
AverageIntoBins
}
\details{
AverageIntoBins
}
\examples{
\dontrun{
}
}
\author{
Juan Inda, <inda@chalmers.se>
}
