% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc2t_binningInfo.R
\docType{data}
\name{bc2t_binningInfo}
\alias{bc2t_binningInfo}
\title{bc2t_binningInfo data set (NAME)}
\format{Is it a matrix?}
\source{
Where is it coming from?
}
\usage{
bc2t_binningInfo
}
\description{
Description 1
}
\details{
Description 2
}
\keyword{datasets}
