% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc2t_UMIdistribution.R
\docType{data}
\name{bc2t_UMIdistribution}
\alias{bc2t_UMIdistribution}
\title{bc2t_UMIdistribution}
\format{Double-precision vector}
\source{
Calculated from the BC dataset.
}
\usage{
bc2t_UMIdistribution
}
\description{
Part of the DSAVE standard template. Describes number of counts for each cell in the
down-sampled dataset after alignment.
}
\keyword{datasets}
