% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binInf2000.R
\docType{data}
\name{binInf2000}
\alias{binInf2000}
\title{binInf2000}
\format{
List with three fields: binninginfo.means, binninginfo.ubs and binninginfo.lbs,
corresponding to the mean, upper bound and lower bound for each bin. Each of those is a
list of 1,000 points.
}
\source{
Calculated from the BC dataset
}
\usage{
binInf2000
}
\description{
Part of the DSAVE standard template. Describes the edges for the bins to use
for calculating variation per gene expression.
}
\keyword{datasets}
