% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkTotalVar1vs1.R
\docType{data}
\name{bulkTotalVar1vs1}
\alias{bulkTotalVar1vs1}
\title{bulkTotalVar1vs1 data set (NAME)}
\format{Is it a matrix?}
\source{
Where is it coming from?
}
\usage{
bulkTotalVar1vs1
}
\description{
Description 1
}
\details{
Description 2
}
\keyword{datasets}
