% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkTotalVar4v4.R
\docType{data}
\name{bulkTotalVar4vs4}
\alias{bulkTotalVar4vs4}
\title{bulkTotalVar4vs4 data set (NAME)}
\format{Is it a matrix?}
\source{
Where is it coming from?
}
\usage{
bulkTotalVar4vs4
}
\description{
Description 1
}
\details{
Description 2
}
\keyword{datasets}
