% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPackageVariables.R
\name{createPackageVariables}
\alias{createPackageVariables}
\title{createPackageVariables}
\usage{
createPackageVariables()
}
\description{
Sets up all package data except the standard template, which was done manually.
Make sure the current directory is the package root before calling this function.
}
\author{
Johan Gustafsson, <gustajo@chalmers.se>
}
