% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetScoresHuman.R
\docType{data}
\name{datasetScoresHuman}
\alias{datasetScoresHuman}
\title{datasetScoresHuman data set (NAME)}
\format{Is it a matrix?}
\source{
Where is it coming from?
}
\usage{
datasetScoresHuman
}
\description{
Description 1
}
\details{
Description 2
}
\keyword{datasets}
