% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVE.R
\docType{package}
\name{DSAVE}
\alias{DSAVE}
\alias{DSAVE-package}
\title{DSAVE}
\description{
Dissecting Cell-to-Cell Variation in Single-Cell RNA-Seq Data
}
\details{
R Package DSAVE, Down-SAmpling based Variation Estimation. The package is useful for comparing the variation that does not originate from sampling, defined as BTM variation, between cell populations, including across datasets, as a quality metric.
}
\author{
Juan Inda, <inda@chalmers.se>, Johan Gustafsson, <gustajo@chalmers.se>
}
