% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVEAlignDataset.R
\name{DSAVEAlignDataset}
\alias{DSAVEAlignDataset}
\title{DSAVEAlignDataset}
\usage{
DSAVEAlignDataset(data, templInfo)
}
\arguments{
\item{data}{matrix or sparse matrix, the type is preserved in the output}

\item{templInfo}{list with one element called UMIDistr containing template information}
}
\value{
a matrix
}
\description{
Align a cell population to a template
}
\details{
Aligns the cell population to the template, by removing genes, removing cells and down-sampling the
data to match the template as good as possible.
All cell populations successfully aligned to the same template will have almost
identical sampling noise.
}
\examples{
dsAligned = DSAVEAlignDataset(data, templInfo)

}
\author{
Juan Inda, <inda@chalmers.se>
}
