% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVECalcBTMScore.R
\name{DSAVECalcBTMScore}
\alias{DSAVECalcBTMScore}
\title{DSAVECalcBTMScore}
\usage{
DSAVECalcBTMScore(data, templInfo, skipAlignment = FALSE,
  iterations = 15, useLogTransform = FALSE, logTPMAddon = 1,
  silent = FALSE)
}
\arguments{
\item{data}{numeric matrix (can be sparse), the input dataset (cell population)}

\item{templInfo}{template information}

\item{skipAlignment}{logical, default FALSE. If FALSE, alignment will be done against the template data}

\item{iterations}{number of iterations; defaults to 15}

\item{useLogTransform}{default FALSE, if TRUE calculation will be done on log transformed data}

\item{logTPMAddon}{default 1, when using log transformed data, this number is added to the data before transforming to avoid log(0)}

\item{silent}{(optional) If true, no progress bar is shown. Defaults to FALSE}
}
\value{
list
}
\description{
Calculate BTM Score
}
\details{
Calculate BTM Score
}
\examples{
\dontrun{
}
}
\author{
Juan Inda, <inda@chalmers.se>
}
