% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVEDivParamPlot.R
\name{DSAVEDivParamPlot}
\alias{DSAVEDivParamPlot}
\title{DSAVEDivParamPlot}
\usage{
DSAVEDivParamPlot(divData, paramData, paramName = "Param")
}
\arguments{
\item{divData}{The output from the function DSAVEGetSingleCellDivergence}

\item{paramData}{Vector containing the param value for each cell.}
}
\value{
-
}
\description{
Shows number of UMIs as function of divergence. Useful for determining thresholds for discarding bad cells.
}
\examples{
\dontrun{DSAVEDivParamPlot(data, divData)}

}
\author{
Johan Gustafsson, <gustajo@chalmers.se>
}
