% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVEDivUMIPlot.R
\name{DSAVEDivUMIPlot}
\alias{DSAVEDivUMIPlot}
\title{DSAVEDivUMIPlot}
\usage{
DSAVEDivUMIPlot(data, divData)
}
\arguments{
\item{data}{The UMI count data}

\item{divData}{The output from the function DSAVEGetSingleCellDivergence}
}
\value{
nothing
}
\description{
Shows number of UMIs as function of divergence. Useful for determining thresholds for discarding bad cells.
}
\examples{
\dontrun{DSAVEDivUMIPlot(data, divData)}

}
\author{
Johan Gustafsson, <gustajo@chalmers.se>
}
