% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVEGenerateTemplate.R
\name{DSAVEGenerateTemplate}
\alias{DSAVEGenerateTemplate}
\title{DSAVEGenerateTemplate}
\usage{
DSAVEGenerateTemplate(
  ds,
  genesToUse = NULL,
  numCells = 2000,
  numUMIs = 750,
  fractionUpperOutliers = 0.025,
  fractionLowerOutliers = 0.025
)
}
\arguments{
\item{ds}{a dataset, numerical matrix with genes as row names}

\item{genesToUse}{(optional) List of genes to filter on in the template; defaults to the genes in ds}

\item{numCells}{(optional) Number of cells to include in the template. Defaults to 2000.}

\item{numUMIs}{(optional)  Average number of UMIs per cell to aim for in the template
Defaults to 750.}

\item{fractionUpperOutliers}{the fraction of upper outliers that should be discarded.
Defaults to 0.025.}

\item{fractionLowerOutliers}{the fraction of lower outliers that should be discarded.
Defaults to 0.025.}
}
\value{
list<...>
}
\description{
Creates a template from an existing dataset and a gene list
}
\author{
Johan Gustafsson, <gustajo@chalmers.se>
}
