% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVEGetGeneVariation.R
\name{DSAVEGetGeneVariation}
\alias{DSAVEGetGeneVariation}
\title{DSAVEGetGeneVariation}
\usage{
DSAVEGetGeneVariation(data, lb = 10, iterations = 100,
  maxNumCells = 2000, silent = FALSE)
}
\arguments{
\item{data}{numeric matrix, the input dataset (cell population)}

\item{lb}{(optional) TPM lower bound, genes below this will not be investigated. Defaults to 10 TPM/CPM.}

\item{iterations}{(optional) The number of SNO datasets to generate.
Recommended value is 100 if no p values are needed,
10,000 - 100,000 if p-values are of interest. Defaults to 100,000.}

\item{maxNumCells}{(optional) ds is reduced to this number of cells if it
contains more, to save computation time. Defaults to 2,000.}

\item{silent}{(optional) If true, no progress bar is shown. Defaults to FALSE}
}
\value{
list(genes, logCVDifference, pVals, SNOVariances, SNOCountsPerGene)
}
\description{
Calculates the DSAVE gene-wise BTM variation metric.
}
\author{
Juan Inda, <inda@chalmers.se>, Johan Gustafsson, <gustajo@chalmers.se>
}
