% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVEGetSingleCellDivergence.R
\name{DSAVEGetSingleCellDivergence}
\alias{DSAVEGetSingleCellDivergence}
\title{DSAVEGetSingleCellDivergence}
\usage{
DSAVEGetSingleCellDivergence(data, minUMIsPerCell = 200,
  tpmLowerBound = 0, iterations = 15, silent = FALSE)
}
\arguments{
\item{data}{numeric matrix (can be sparse), the input dataset (cell population)}

\item{minUMIsPerCell}{All cells are downsampled to this number of counts, or.
higher if possible without losing cells. The result is NA for cells with lower counts.}

\item{tpmLowerBound}{(optional) TPM lower bound, genes below this will not
be included. Defaults to 0 (meaning all genes are included).}

\item{iterations}{(optional) The number of times to iterate. Defaults to 15.}

\item{silent}{(optional) If true, no progress bar is shown. Defaults to FALSE}
}
\value{
vector of divergence, one value per cell.
}
\description{
Calculates the DSAVE cell-wise variation metric.
}
\details{
The divergence is the
log-likelihood for getting the observed counts' distribution for each
cell when sampling counts from the mean dataset gene expression. The
values are negative, and the lower (i.e. more negative) the value, the
more divergent the cell is.
}
\examples{
\dontrun{a = DSAVEGetSingleCellDivergence(data)}
}
\author{
Johan Gustafsson, <gustajo@chalmers.se>
}
