% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVEGetStandardTemplate.R
\name{DSAVEGetStandardTemplate1000}
\alias{DSAVEGetStandardTemplate1000}
\title{DSAVEGetStandardTemplate1000}
\usage{
DSAVEGetStandardTemplate1000()
}
\value{
the template
}
\description{
Gets the standard template that is distributed with the package, requiring > 1000 cells.
}
\author{
Johan Gustafsson, <gustajo@chalmers.se>
}
