% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVEGetStandardTemplate.R
\name{DSAVEGetStandardTemplate500}
\alias{DSAVEGetStandardTemplate500}
\title{DSAVEGetStandardTemplate500}
\usage{
DSAVEGetStandardTemplate500()
}
\value{
the template
}
\description{
Gets the standard template that is distributed with the package, requiring > 500 cells.
}
\author{
Johan Gustafsson, <gustajo@chalmers.se>
}
