% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVEGetTotalVariationFromBulk.R
\name{DSAVEGetTotalVariationFromBulk}
\alias{DSAVEGetTotalVariationFromBulk}
\title{DSAVEGetTotalVariationFromBulk}
\usage{
DSAVEGetTotalVariationFromBulk(data, pool4samples, upperBound = 1e+05,
  lowerBound = 0.5, na.rm = TRUE, nComb = 1000L, rescale = TRUE)
}
\arguments{
\item{data}{a numericl matrix}

\item{pool4samples}{logic value, if TRUE, the function will compare the mean
of 4 randomly selected samples with the mean of 4 others}

\item{upperBound}{filters out the genes with mean expression higher than this value}

\item{lowerBound}{filters out the genes with mean expression lower than this value}

\item{na.rm}{logical, sends na.rm value to the rowMeans function}

\item{nComb}{?}

\item{rescale}{logical, determine if the data should be rescaled to TPM}
}
\value{
numeric vector
}
\description{
Calculate total variation from bulk sample
}
\details{
Calculates the average pairwise total variation between a list of bulk samples, with a TPM
filtration on the genes. If pool4samples is true, it will compare the mean
of 4 randomly selected samples with the mean of 4 others. This expects a
list of 8 samples.
}
\examples{
\dontrun{
DSAVEGetTotalVariationFromBulk(sample, pool4samples = FALSE, upperBound = 100000,
lowerBound = 0.5)
}

}
\author{
Juan Inda, <inda@chalmers.se>
}
