% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVEPlotDSAVEScore.R
\name{DSAVEPlotDSAVEScore}
\alias{DSAVEPlotDSAVEScore}
\title{DSAVEPlotDSAVEScore}
\usage{
DSAVEPlotDSAVEScore(dsScores, dsNames, ref = NA)
}
\arguments{
\item{dsScores}{Output from DSAVEGetTotalVariation or a list of such outputs}

\item{dsNames}{The names of the dataset(s), either a string or a list of strings}

\item{ref}{(Optional) Reference datasets, for example DSAVE::datasetScoresHuman (2000),
DSAVE::datasetScoresHuman1000 and DSAVE::datasetScoresHuman500. The reference must have been
calculated using the same template as the samples, i.e. match the number of cells.}
}
\value{
nothing
}
\description{
Plots DSAVE scores, optionally with scores from other datasets as reference.
}
\details{
Calculates the average pairwise total variation between two pools of cells with specific size,
with a TPM filtration on the genes.
}
\examples{
\dontrun{DSAVEPlotDSAVEScore(dsScores, dsNames, DSAVE::datasetScoresHuman)}

}
\author{
Johan Gustafsson, <gustajo@chalmers.se>
}
