% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVEPlotTotalVariation.R
\name{DSAVEPlotTotalVariation}
\alias{DSAVEPlotTotalVariation}
\title{DSAVEPlotTotalVariation}
\usage{
DSAVEPlotTotalVariation(dsData, dsNames)
}
\arguments{
\item{dsData}{Output from DSAVEGetTotalVariation or a list of such outputs}

\item{dsNames}{The names of the dataset(s), either a string or a list of strings}
}
\value{
nothing
}
\description{
Plots the total variation as a function of pool size, with bulk variation as reference.
}
\examples{
\dontrun{DSAVEPlotTotalVariation(dsData, dsNames)}

}
\author{
Johan Gustafsson, <gustajo@chalmers.se>
}
