% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSeuratData.R
\name{extractSeuratData}
\alias{extractSeuratData}
\title{extractSeuratData}
\usage{
extractSeuratData(seuratObj, clusterId = NA)
}
\value{
the template
}
\description{
Extracts the data from a Seurat object. Optionally, a cluster id can be sent in, to extract
a subset of the data
}
\author{
Johan Gustafsson, <gustajo@chalmers.se>
}
