% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBinningInfo.R
\name{generateBinningInfo}
\alias{generateBinningInfo}
\title{generateBinningInfo}
\usage{
generateBinningInfo(data, lbnonlog = 5L, ubnonlog = 1300L,
  rescale = TRUE, numPoints = 1000L, poolSize = 500L, step = 5e-04,
  toplim = 1000L)
}
\arguments{
\item{data}{Numeric matrix}

\item{lbnonlog}{lower bound in original scale. Should stretch a bit outside the intended range.}

\item{ubnonlog}{upper bound in original scale. Should stretch a bit outside the intended range.}

\item{rescale}{If the data should be rescaled to TPM before the calculation. Defaults to true.}

\item{numPoints}{number of points for the distribution}

\item{poolSize}{number of genes to pool at each point in the distribution}

\item{step}{increase in lower and upper bound to get enough genes at each point in the distribution}

\item{toplim}{maximum number of iterations to obtain the frequency at each point in the distribution}
}
\value{
list(means, lower bounds, upper bounds)
}
\description{
Creates the part of the template called the binning information from a dataset.
}
\author{
Juan Inda, <inda@chalmers.se>
}
