% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateUMIdistribution.R
\name{generateUMIDistribution}
\alias{generateUMIDistribution}
\title{generateUMIDistribution}
\usage{
generateUMIDistribution(data, numUMI = 750)
}
\arguments{
\item{data}{Numeric matrix}

\item{numUMI}{average number of molecules per cell}
}
\value{
a template object
}
\description{
Creates the part of the template called the UMI distribution from a dataset.
}
\author{
Juan Inda, <inda@chalmers.se>
}
