% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesForTemplate.R
\docType{data}
\name{genesForTemplate}
\alias{genesForTemplate}
\title{genesForTemplate}
\format{string list}
\source{
The genes are the intersection of the genes available in four datasets; BC, OC, B10k and TCD8.
}
\usage{
genesForTemplate
}
\description{
Part of the standard template. Intersection of genes found in 4 datasets of single cell RNA-seq.
}
\keyword{datasets}
