% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenSampDs.R
\name{GenSampDs}
\alias{GenSampDs}
\title{GenSampDs}
\usage{
GenSampDs(data, countsPerGene, edges)
}
\arguments{
\item{data}{numeric matrix, the input dataset (cell population)}

\item{countsPerGene}{counts per gene}

\item{edges}{edges}
}
\value{
list(logCV, variances)
}
\description{
GenSampDs
}
\details{
GenSampDs
}
\examples{
\dontrun{
}

}
\author{
Juan Inda, <inda@chalmers.se>
}
