% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGeneCVs.R
\name{GetGeneCVs}
\alias{GetGeneCVs}
\title{GetGeneCVs}
\usage{
GetGeneCVs(ds, logTPMAddon, toLog = FALSE)
}
\arguments{
\item{ds}{numeric matrix}

\item{logTPMAddon}{default 1, when using log transformed data, this number is added to the data before transforming to avoid log(0)}

\item{toLog}{default FALSE, if TRUE calculation will be done on log transformed data}
}
\value{
a vector
}
\description{
GetGeneCVs
}
\details{
Calculate coefficient of variation
}
\examples{
\dontrun{
}
}
\author{
Juan Inda, <inda@chalmers.se>
}
