% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLogFactorials.R
\name{GetLogFactorials}
\alias{GetLogFactorials}
\title{GetLogFactorials}
\usage{
GetLogFactorials(n)
}
\arguments{
\item{n}{numeric, the max value needed}
}
\value{
vector
}
\description{
GetLogFactorials
}
\details{
Calculates the logarithm of x! for all x <= n and places those in a vector.
  To get around that you cannot index a vector with 0, you should always
  index this vector with the value + 1.
  The purpose of this function is that x! cannot be calculated for
  x > 170, while that is not a problem for log(x!), but you cannot use
  the built-in factorial function in matlab and do log afterwards!
}
\author{
Johan Gustafsson, Juan Inda, <inda@chalmers.se>
}
