% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetVarAndLogCV.R
\name{GetVarAndLogCV}
\alias{GetVarAndLogCV}
\title{GetVarAndLogCV}
\usage{
GetVarAndLogCV(dsdata, UMIsPerCell)
}
\arguments{
\item{dsdata}{numeric matrix, the input dataset (cell population)}

\item{UMIsPerCell}{number of UMIs to normalize by.}
}
\value{
list(logCV, variances)
}
\description{
Help function for DSAVEGetGeneVariation. Calculates variation and log CV.
}
\details{
GetVarAndLogCV
}
\examples{
\dontrun{
}
}
\author{
Juan Inda, <inda@chalmers.se>
}
