% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportTabSepSamples.R
\name{ImportTabSepSamples}
\alias{ImportTabSepSamples}
\title{Text file to Sample object}
\usage{
ImportTabSepSamples(filename, sampleName)
}
\arguments{
\item{filename}{path to the file}

\item{sampleName}{name of the Sample object}
}
\description{
Imports a txt file and creates a Sample object
}
\details{
Imports a txt tab separated, takes first column as rows' names and first row as columns' name
}
\examples{
\dontrun{
samp <- ImportTabSepSamples("~/Downloads/dsaver/tcellCD4ProfilesTMMNormalized.txt",
"tcellCD4Profiles")
}
}
\author{
Juan Inda, <inda@chalmers.se>
}
