% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadAndDownloadBCells.R
\name{LoadAndDownloadBCells}
\alias{LoadAndDownloadBCells}
\title{LoadAndDownloadBCells}
\usage{
LoadAndDownloadBCells()
}
\value{
the template
}
\description{
Gets the standard template that is redistributed with the package. This function uses Seurat.
You need to run the following two lines to use it:
install.packages('Seurat')
library(Seurat)
}
\examples{
\dontrun{ LoadAndDownloadBCells()
}
}
\author{
Johan Gustafsson, <gustajo@chalmers.se>
}
