% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogBinomialPDF.R
\name{LogBinomialPDF}
\alias{LogBinomialPDF}
\title{LogBinomialPDF}
\usage{
LogBinomialPDF(obs, prob, logFacs = NULL)
}
\arguments{
\item{obs}{numeric, observed values}

\item{prob}{numeric, probabilities}

\item{logFacs}{(optional) Precalculated vertical vector of log factorials,
created using the function GetLogFactorials(n),
where n is at least the sum of all observations.}
}
\value{
vector

\dontrun{LogBinomialPDF(obs, prob)}
}
\description{
LogBinomialPDF
}
\details{
Calculates the probability density function for a vector of observations
  for a binomial distribution.
  The idea with this function is that it is not easy to calculate the PDF of
  a binomial if n is large, since we cannot practically calculate n! of a
  large number. The logarithm of the pdf is fine to calculate however, since
  log(n!) = log(n) + log(n-1) + ... + log(1). So we cannot use the
  built-in function for the pdf and afterwards log it, but need to implement
  the function ourselves instead.
}
\author{
Johan Gustafsson, Juan Inda, <inda@chalmers.se>
}
