% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBMC68K_Tcells.R
\docType{data}
\name{PBMC68K_Tcells}
\alias{PBMC68K_Tcells}
\title{PBMC68K_Tcells data set}
\format{A Sample object}
\source{
10X single cell data from...
}
\usage{
PBMC68K_Tcells
}
\description{
T cells from the PBMC68000 data set
}
\details{
A randdom subsample of 8000 cells from.
T cells were obtained from the whole data set as they had been previously classified
by the authors.
}
\keyword{datasets}
