% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Samples.R
\docType{data}
\name{Samples}
\alias{Samples}
\title{Samples object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Samples
}
\arguments{
\item{name}{Name of the object}

\item{data}{Numeric matrix}

\item{genes}{name of the rows or genes}

\item{sampleIds}{name of the columns or samples}

\item{genesToKeep}{subsample of the genes to keep}

\item{sortGenes}{.}

\item{name}{for inner and outerjoin the name of the added Sample object if it was a matrix}

\item{ds}{another sample object or matrix to do merge to the original sample}
}
\value{
a Sample object
}
\description{
Samples object elements and methods
}
\details{
Creates Samples object, updates its values and contains the functions of
this object
}
\examples{
\dontrun{
s <- Samples$new("tcellCD4Profiles", tcellCD4Profiles)
s2 <- s$geneSubset(genesToKeep = rownames(tcellCD4Profiles)[1:10])
s3 <- s$geneSubset(genesToKeep = rownames(tcellCD4Profiles)[11:20])
tpm1 <- s$innerJoin(s2, "s2")
colnames(s3$data) <- paste(colnames(s3$data),"s3")
tpm2 <- s2$fullOuterJoin(s3)
}
}
\author{
Juan Inda, <inda@chalmers.se>
}
\keyword{datasets}
