% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBinningInfo.R
\name{setBinningInfo}
\alias{setBinningInfo}
\title{BINNING INFORMATION}
\usage{
setBinningInfo(data = NULL, lbnonlog = NULL, ubnonlog = NULL,
  reScale = FALSE, numPoints = 1000L, poolSize = 500L,
  step = 5e-04, toplim = 1000L)
}
\arguments{
\item{data}{Numeric matrix}

\item{lbnonlog}{lower bound in original scale}

\item{ubnonlog}{upper bound in original scale}

\item{reScale}{maximum number of iterations to obtain the frequency at each point in the distribution}

\item{numPoints}{number of points for the distribution}

\item{poolSize}{number of genes to pool at each point in the distribution}

\item{step}{increas in lower and upper bound to get enough genes at each point in the distribution}

\item{toplim}{maximum number of iterations to obtain the frequency at each point in the distribution}
}
\value{
a template object
}
\description{
Samples object elements and methods
}
\details{
Sets the binning information
}
\examples{
\dontrun{
}
}
\author{
Juan Inda, <inda@chalmers.se>
}
