% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchronizeMatrices.R
\name{synchronizeMatrices}
\alias{synchronizeMatrices}
\title{synchronizeMatrices}
\usage{
synchronizeMatrices(mat1 = NULL, mat2 = NULL, discardRows = TRUE)
}
\arguments{
\item{mat1}{A matrix or vector}

\item{mat2}{A matrix or vector}

\item{discardRows}{a logic vector, TRUE for intersect, FALSE for union}
}
\value{
A matrix or vector
}
\description{
Not so optimal merge of vectors or matrices
}
\details{
Merges vectors or matrices, or vectors and matrices
}
\author{
Juan Inda, <inda@chalmers.se>
}
