% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcellCD4ProfilesTMMNormalized.R
\docType{data}
\name{tcellCD4ProfilesTMMNormalized}
\alias{tcellCD4ProfilesTMMNormalized}
\title{tcellCD4ProfilesTMMNormalized data set}
\format{A Sample object}
\source{
bulkd data from...
}
\usage{
tcellCD4ProfilesTMMNormalized
}
\description{
Bulk T cells data
}
\details{
8 bulk T cells data
}
\keyword{datasets}
