% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templInfo.R
\docType{data}
\name{templInfo}
\alias{templInfo}
\title{Template object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
templInfo
}
\arguments{
\item{name}{Name of the object}

\item{data}{Numeric matrix}

\item{genesToKeep}{genes to include in the template}

\item{numCells}{number of cells in the template}

\item{numUMI}{average number of molecules per cell}

\item{fractionUpperOutliers}{fraction of upper outliers}

\item{fractionLowerOutliers}{fraction of lower outliers}

\item{binningInfo.lbs}{distribution}

\item{binningInfo.ubs}{distribution}

\item{UMIDistr}{distribution of UMIs}

\item{template.built}{logical}

\item{lbnonlog}{lower bound in original scale}

\item{ubnonlog}{upper bound in original scale}

\item{numPoints}{number of points for the distribution}

\item{poolSize}{number of genes to pool at each point in the distribution}

\item{step}{increas in lower and upper bound to get enough genes at each point in the distribution}

\item{toplim}{maximum number of iterations to obtain the frequency at each point in the distribution}
}
\value{
a template object
}
\description{
Samples object elements and methods
}
\details{
Creates Samples object, updates its values and contains the functions of
this object
}
\examples{
\dontrun{
}
}
\author{
Juan Inda, <inda@chalmers.se>
}
\keyword{datasets}
