% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpmDSAVE.R
\name{tpmDSAVE}
\alias{tpmDSAVE}
\title{tpmDSAVE}
\usage{
tpmDSAVE(counts)
}
\arguments{
\item{count}{A matrix.}
}
\value{
A matrix
}
\description{
Standarizes by column making it sum to 1e6
}
\details{
Standarizes by column making it sum to 1e6
}
\author{
Juan Inda, <inda@chalmers.se>
}
