% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMIDistr2000.R
\docType{data}
\name{UMIDistr2000}
\alias{UMIDistr2000}
\title{UMIDistr2000}
\format{Double-precision vector}
\source{
Calculated from the BC dataset.
}
\usage{
UMIDistr2000
}
\description{
Part of the DSAVE standard template. Describes number of counts for each cell in the
down-sampled dataset after alignment.
}
\keyword{datasets}
