% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageIntoBins.R
\name{averageIntoBins}
\alias{averageIntoBins}
\title{averageIntoBins}
\usage{
averageIntoBins(TPMData, logcv, templInfo)
}
\arguments{
\item{TPMData}{numeric matrix, tpmdata}

\item{logcv}{numeric vector, Gene Coefficient of Variation}

\item{templInfo}{list with 5 elements}
}
\value{
a list with CV and meanGeneExpr
}
\description{
Help function for the DSAVE Score calculation. Places genes into
gene expression range bins and then calculates the mean for each
bin.
}
\details{
averageIntoBins
}
\author{
Juan Inda, <inda@chalmers.se>
}
