% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc2t_UMIdistribution.R
\docType{data}
\name{bc2t_UMIdistribution}
\alias{bc2t_UMIdistribution}
\title{bc2t_UMIdistribution data set (NAME)}
\format{Is it a matrix?}
\source{
Where is it coming from?
}
\usage{
bc2t_UMIdistribution
}
\description{
Description 1
}
\details{
Description 2
}
\keyword{datasets}
