% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkTotalVar4v4.R, R/bulkTotalVar4vs4.R
\docType{data}
\name{bulkTotalVar4vs4}
\alias{bulkTotalVar4vs4}
\title{bulkTotalVar4vs4 data set (NAME)}
\format{Is it a matrix?}
\source{
Where is it coming from?

We downloaded 8 bulk RNA-Seq samples (FPKM) from the BLUEPRINT Epigenome Project31. 
The samples were taken from the project EGAD00001001173 and have the following sample IDs: 
S002EV11, S004M711, S007DD11, S007G711, S008H111, S009W411, S0018A13, and S0041C11. 
The FPKM data were normalized using TMM32 before use, and scaled to an average count of 10^6 per sample.
}
\usage{
bulkTotalVar4vs4

bulkTotalVar4vs4
}
\description{
Description 1

Average total variation between bulk samples at different expression intervals (similar to TPM).
The samples are first averaged four by four, and then compared.
}
\details{
Description 2
}
\keyword{datasets}
