% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPackageVariables.R
\name{createPackageVariables}
\alias{createPackageVariables}
\title{LoadAndDownloadT4k}
\usage{
createPackageVariables()
}
\value{
nothing
}
\description{
Sets up all package data except the standard template, which was done manually
Make sure the current directory is the package root
}
\examples{
\dontrun{ createPackageVariables()
}
}
\author{
Johan Gustafsson, <gustajo@chalmers.se>
}
