% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVE.R
\docType{package}
\name{DSAVE}
\alias{DSAVE}
\alias{DSAVE-package}
\title{DSAVE}
\description{
Dissecting Cell-to-Cell Variation in Single-Cell RNA-Seq Data
}
\details{
R Package DSAVE, Down-SAmpling based Variation Estimation. The package is useful for comparing the variation in single-cell RNA-Seq data
that does not originate from sampling, defined as BTM variation, between cell populations. The package supports
comparing variation across datasets and can be used as a quality metric, where the variation can be compared
to that of other datasets from similar conditions.
}
\author{
Johan Gustafsson, <johan.gson@gmail.com>, Juan Inda, <inda@chalmers.se>
}
