% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVECalcBTMScore.R
\name{DSAVECalcBTMScore}
\alias{DSAVECalcBTMScore}
\title{DSAVECalcBTMScore}
\usage{
DSAVECalcBTMScore(data, templInfo, skipAlignment = FALSE,
  iterations = 15, useLogTransform = FALSE, logTPMAddon = 1,
  silent = FALSE)
}
\arguments{
\item{data}{numeric matrix (can be sparse), the input dataset (cell population)}

\item{templInfo}{template information}

\item{skipAlignment}{logical, default FALSE. If FALSE, alignment will be done against the template data}

\item{iterations}{number of iterations; defaults to 15}

\item{useLogTransform}{default FALSE, if TRUE calculation will be done on log transformed data}

\item{logTPMAddon}{default 1, when using log transformed data, this number is added to the data before transforming to avoid log(0)}

\item{silent}{(optional) If true, no progress bar is shown. Defaults to FALSE}
}
\value{
List containing total variation from the aligned and SNO cell populations over the gene expression range,
the difference between those, the corresponding gene expression values, and finally the BTM score.
}
\description{
Calculates the DSAVE BTM Variation Score, which gives and overall score for the BTM variation across all genes.
This metric is comparable across datasets.
}
\details{
The function generates sampling noise only (SNO) datasets for comparison with the real data,
and subtracts the SNO total variation (which is sampling noise only) from the total variation of the
cell population, yielding the BTM variation. Cell population alignment makes the metric comparable
across cell populations and datasets.
}
\examples{
a = DSAVECalcBTMScore(data, templInfo)

}
\author{
Juan Inda, <inda@chalmers.se>, Johan Gustafsson, <gustajo@chalmers.se>
}
