% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVEGetStandardTemplate.R
\name{DSAVEGetStandardTemplate}
\alias{DSAVEGetStandardTemplate}
\title{DSAVEGetStandardTemplate}
\usage{
DSAVEGetStandardTemplate()
}
\value{
the template
}
\description{
Gets the standard template that is distributed with the package.
}
\author{
Johan Gustafsson, <gustajo@chalmers.se>
}
