% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVEGetStandardTemplate.R
\name{DSAVEGetStandardTemplate}
\alias{DSAVEGetStandardTemplate}
\title{DSAVEGetStandardTemplate}
\usage{
DSAVEGetStandardTemplate()
}
\value{
the template
}
\description{
Gets the standard template that is distributed with the package, requiring > 2000 cells.
There are variants for 2000 (standard), 1000 and 500 cells.
}
\author{
Johan Gustafsson, <gustajo@chalmers.se>
}
