% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVEGetTotalVariationPoolSize.R
\name{DSAVEGetTotalVariationPoolSize}
\alias{DSAVEGetTotalVariationPoolSize}
\title{DSAVEGetTotalVariationPoolSize}
\usage{
DSAVEGetTotalVariationPoolSize(data, poolSize = 4,
  upperBoundTPM = 1e+05, lowerBoundTPM = 0.5, na.rm = TRUE,
  seed = NULL, repetitionPerSize = 30L, rescale = TRUE)
}
\arguments{
\item{data}{a numeric matrix}

\item{poolSize}{numeric value, it should be maximum half of the total size of the data
in the data object}

\item{upperBoundTPM}{filters out the genes with mean expression higher than this value}

\item{lowerBoundTPM}{filters out the genes with mean expression lower than this value}

\item{na.rm}{logical, sends na.rm value to the rowMeans function}

\item{seed}{positive integer for to set the seed of the datas}

\item{repetitionPerSize}{positive integer how many times the variation is calculated at each
pool size}

\item{rescale}{logical, determine if the data should be rescaled}
}
\value{
numeric vector
}
\description{
Calculate total variation with a specific pool size
}
\details{
Calculates the average pairwise total variation between two pools of cells with specific size,
with a TPM filtration on the genes.
}
\examples{
\dontrun{
DSAVEGetTotalVariationPoolSize(data, poolSize = 50, upperBoundTPM = 1000,
lowerBoundTPM = 0.01, na.rm = TRUE, seed = 1, repetitionPerSize = 30, 0.5)
}

}
\author{
Juan Inda, <inda@chalmers.se>
}
