% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVEPlotDivergence.R
\name{DSAVEPlotDivergence}
\alias{DSAVEPlotDivergence}
\title{DSAVEPlotDivergence}
\usage{
DSAVEPlotDivergence(data, divData)
}
\arguments{
\item{data}{The UMI count data}

\item{divData}{The output from the function DSAVEGetSingleCellDivergence}
}
\value{
nothing
}
\description{
Calculates the total variation with a specific pool size.
}
\details{
Calculates the average pairwise total variation between two pools of cells with specific size,
with a TPM filtration on the genes.
}
\examples{
\dontrun{DSAVEPlotDivergence(data, divData)}

}
\author{
Johan Gustafsson, <gustajo@chalmers.se>
}
