% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAVEPlotTotalVariation.R
\name{DSAVEPlotTotalVariation}
\alias{DSAVEPlotTotalVariation}
\title{DSAVEPlotTotalVariation}
\usage{
DSAVEPlotTotalVariation(dsData, dsNames, bulkIndex = 1)
}
\arguments{
\item{dsData}{Output from DSAVEGetTotalVariation or a list of such outputs}

\item{dsNames}{The names of the dataset(s), either a string or a list of strings}

\item{[optional]}{bulkIndex A number 1-5, representing the following pseudo-TPM values: 1: 0.5-100k,
2: 0.5-2, 3: 0.5-50, 4: 0.5-50, 5: 2-100. Defaults to 1.}
}
\value{
the ggplot object
}
\description{
Plots the total variation as a function of pool size, with bulk variation as reference.
}
\examples{
\dontrun{DSAVEPlotTotalVariation(dsData, dsNames)}

}
\author{
Johan Gustafsson, <gustajo@chalmers.se>
}
