% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesForTemplate.R
\docType{data}
\name{genesForTemplate}
\alias{genesForTemplate}
\title{genesForTemplate}
\format{A Sample object}
\source{
10X single cell data from...
}
\usage{
genesForTemplate
}
\description{
Intersection of genes found in 4 datasets of single cell RNA-seq
}
\details{
Intersection of genes found in 4 datasets of single cell RNA-seq
}
\keyword{datasets}
