% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetVarAndLogCV.R
\name{GetVarAndLogCV}
\alias{GetVarAndLogCV}
\title{GetVarAndLogCV}
\usage{
GetVarAndLogCV(dsdata, UMIsPerCell)
}
\arguments{
\item{UMIsPerCell}{number of UMIs to normalize by.}

\item{data}{numeric matrix, the input dataset (cell population)}
}
\value{
list(logCV, variances)
}
\description{
GetVarAndLogCV
}
\details{
GetVarAndLogCV
}
\examples{
\dontrun{
}
}
\author{
Juan Inda, <inda@chalmers.se>
}
