% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarAndLogCV.R
\name{getVarAndLogCV}
\alias{getVarAndLogCV}
\title{getVarAndLogCV}
\usage{
getVarAndLogCV(dsdata, UMIsPerCell)
}
\arguments{
\item{dsdata}{numeric matrix, the input dataset (cell population)}

\item{UMIsPerCell}{number of UMIs to normalize by.}
}
\value{
list(logCV, variances)
}
\description{
Help function for DSAVEGetGeneVariation. Calculates variation and log CV.
}
\details{
getVarAndLogCV
}
\author{
Juan Inda, <inda@chalmers.se>
}
