% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUMIdistribution.R
\name{setUMIDistribution}
\alias{setUMIDistribution}
\title{setUMIDistribution}
\usage{
setUMIDistribution(data, numUMI = 750)
}
\arguments{
\item{data}{Numeric matrix}

\item{numUMI}{average number of molecules per cell}
}
\value{
a template object
}
\description{
Set UMI distribution
}
\details{
Creates a UMI distribution for a template from a given cell population.
}
\author{
Juan Inda, <inda@chalmers.se>
}
